feat: 支持rpc调用sms接口

+++ src/main/java/com/mimiter/model/entity/SMSTemplate.java
@@ -11,7 +11,7 @@ public class SMSTemplate {
 
     private Long templateId;
 
-    private Integer international = 0;  // 是否国际/港澳台短信，其中0表示国内短信，1表示国际/港澳台短信，这里默认为国内短信
+    private Long international = 0L;  // 是否国际/港澳台短信，其中0表示国内短信，1表示国际/港澳台短信，这里默认为国内短信
 
     private Long statusCode;  // 申请模板状态，其中0表示审核通过，1表示审核中，-1表示审核未通过或审核失败
